Only pass InfoRequests to Project::Queue
Now that the backend has been extracted we don't really need the Project
instance encapsulated within the `Project::Queue`.

This is a bit of an intermediary commit – now that we're passing the
`InfoRequest`s directly into the queue we don't need the subclasses at
all.

Replacing all that would make for a huge commit though, so that
will come later.